requirements.txt
black
durable_rules>=2.0.28,<3.0.0
isort